requirements.txt
pytesseract
pillow
reportlab
selenium
webdriver-manager